feat(CreateAcademy.svelte): add functionality to post data to server when form is submitted
The changes in this commit add functionality to the `CreateAcademy.svelte` component to post data to the server when the form is submitted. The `handleSubmit` function now makes an asynchronous call to the `postData` function from the `academy` service, passing in the `name` and `description` values as payload. If the data is successfully posted, a success message is logged to the console and the page is reloaded. If there is an error in posting the data, an error message is logged to the console.