feat(Read.svelte): add support for fetching academy and classroom data for each course
The `loadData` function in `Read.svelte` has been updated to fetch academy and classroom data for each course. This is done by calling the `fetchAcademyById` and `fetchClassroomById` functions from the respective services. The fetched data is stored in `academyMap` and `classroomMap` objects, which are used to display the corresponding names in the table.

The `fetchAcademyById` function has been added to the `academy.ts` service file. It makes a GET request to the `/api/academy/{id}` endpoint to fetch the academy data for a given ID.

The `fetchClassroomById` function has been added to the `classroom.ts` service file. It makes a GET request to the `/api/classroom/{id}` endpoint to fetch the classroom data for a given ID.

These changes were made to improve the user experience by displaying the actual names of the academies and classrooms instead of just their IDs.